Make eval_sql_where available to DefaultPredicateEvaluator
