+++ kernel/src/predicates/parquet_stats_skipping/tests.rs
@@ -1,6 +1,6 @@
 use super::*;
 use crate::expressions::{column_expr, Expression as Expr};
-use crate::predicates::PredicateEvaluator;
+use crate::predicates::PredicateEvaluator as _;
 use crate::DataType;
 
 const TRUE: Option<bool> = Some(true);
@@ -256,127 +256,4 @@ fn test_eval_is_null() {
 
     // all nulls
     do_test(2, &[TRUE, FALSE]);
-}
-
-struct AllNullTestFilter;
-impl ParquetStatsProvider for AllNullTestFilter {
-    fn get_parquet_min_stat(&self, _col: &ColumnName, _data_type: &DataType) -> Option<Scalar> {
-        None
-    }
-
-    fn get_parquet_max_stat(&self, _col: &ColumnName, _data_type: &DataType) -> Option<Scalar> {
-        None
-    }
-
-    fn get_parquet_nullcount_stat(&self, _col: &ColumnName) -> Option<i64> {
-        Some(self.get_parquet_rowcount_stat())
-    }
-
-    fn get_parquet_rowcount_stat(&self) -> i64 {
-        10
-    }
-}
-
-#[test]
-fn test_sql_where() {
-    let col = &column_expr!("x");
-    const VAL: Expr = Expr::Literal(Scalar::Integer(1));
-    const NULL: Expr = Expr::Literal(Scalar::Null(DataType::BOOLEAN));
-    const FALSE: Expr = Expr::Literal(Scalar::Boolean(false));
-    const TRUE: Expr = Expr::Literal(Scalar::Boolean(true));
-
-    // Basic sanity checks
-    expect_eq!(AllNullTestFilter.eval_sql_where(&VAL), None, "WHERE {VAL}");
-    expect_eq!(AllNullTestFilter.eval_sql_where(col), None, "WHERE {col}");
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::is_null(col.clone())),
-        Some(true), // No injected NULL checks
-        "WHERE {col} IS NULL"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::lt(TRUE, FALSE)),
-        Some(false), // Injected NULL checks don't short circuit when inputs are NOT NULL
-        "WHERE {TRUE} < {FALSE}"
-    );
-
-    // Constrast normal vs SQL WHERE semantics - comparison
-    expect_eq!(
-        AllNullTestFilter.eval_expr(&Expr::lt(col.clone(), VAL), false),
-        None,
-        "{col} < {VAL}"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::lt(col.clone(), VAL)),
-        Some(false),
-        "WHERE {col} < {VAL}"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_expr(&Expr::lt(VAL, col.clone()), false),
-        None,
-        "{VAL} < {col}"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::lt(VAL, col.clone())),
-        Some(false),
-        "WHERE {VAL} < {col}"
-    );
-
-    // Constrast normal vs SQL WHERE semantics - comparison inside AND
-    expect_eq!(
-        AllNullTestFilter.eval_expr(&Expr::and(NULL, Expr::lt(col.clone(), VAL)), false),
-        None,
-        "{NULL} AND {col} < {VAL}"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::and(NULL, Expr::lt(col.clone(), VAL),)),
-        Some(false),
-        "WHERE {NULL} AND {col} < {VAL}"
-    );
-
-    expect_eq!(
-        AllNullTestFilter.eval_expr(&Expr::and(TRUE, Expr::lt(col.clone(), VAL)), false),
-        None, // NULL (from the NULL check) is stronger than TRUE
-        "{TRUE} AND {col} < {VAL}"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::and(TRUE, Expr::lt(col.clone(), VAL),)),
-        Some(false), // FALSE (from the NULL check) is stronger than TRUE
-        "WHERE {TRUE} AND {col} < {VAL}"
-    );
-
-    // Contrast normal vs. SQL WHERE semantics - comparison inside AND inside AND
-    expect_eq!(
-        AllNullTestFilter.eval_expr(
-            &Expr::and(TRUE, Expr::and(NULL, Expr::lt(col.clone(), VAL)),),
-            false,
-        ),
-        None,
-        "{TRUE} AND ({NULL} AND {col} < {VAL})"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::and(
-            TRUE,
-            Expr::and(NULL, Expr::lt(col.clone(), VAL)),
-        )),
-        Some(false),
-        "WHERE {TRUE} AND ({NULL} AND {col} < {VAL})"
-    );
-
-    // Semantics are the same for comparison inside OR inside AND
-    expect_eq!(
-        AllNullTestFilter.eval_expr(
-            &Expr::or(FALSE, Expr::and(NULL, Expr::lt(col.clone(), VAL)),),
-            false,
-        ),
-        None,
-        "{FALSE} OR ({NULL} AND {col} < {VAL})"
-    );
-    expect_eq!(
-        AllNullTestFilter.eval_sql_where(&Expr::or(
-            FALSE,
-            Expr::and(NULL, Expr::lt(col.clone(), VAL)),
-        )),
-        None,
-        "WHERE {FALSE} OR ({NULL} AND {col} < {VAL})"
-    );
 }
